Fix warnings; Fix bug in matrix multiply; Add template specialization for 4x4 matrix multiplication

+++ src/core/kraft_application.h
@@ -12,8 +12,8 @@ struct Window;
 
 struct ApplicationConfig
 {
-    size_t              WindowWidth;
-    size_t              WindowHeight;
+    uint32              WindowWidth;
+    uint32              WindowHeight;
     const char*         ApplicationName;
     const char*         WindowTitle;
     RendererBackendType RendererBackend = RendererBackendType::RENDERER_BACKEND_TYPE_NONE;
